requirements.txt
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy @ file:///croot/contourpy_1738160616259/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
easydict==1.13
filelock==3.19.1
fonttools @ file:///home/task_175879159822283/conda-bld/fonttools_1758791613454/work
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
imagecodecs @ file:///home/task_175933549810676/conda-bld/imagecodecs_1759335552109/work
imageio @ file:///croot/imageio_1738159938990/work
importlib_metadata==8.7.0
importlib_resources @ file:///home/task_175940067996674/conda-bld/importlib_resources-suite_1759400795292/work
Jinja2==3.1.6
joblib @ file:///home/task_175792628848965/conda-bld/joblib_1757926345058/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader @ file:///croot/lazy_loader_1718176737906/work
MarkupSafe==3.0.3
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
networkx @ file:///croot/networkx_1717597493534/work
numexpr @ file:///croot/numexpr_1730215937391/work
numpy==1.24.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging @ file:///croot/packaging_1753775359799/work
pandas @ file:///croot/pandas_1756466404410/work/dist/pandas-2.3.2-cp39-cp39-linux_x86_64.whl#sha256=0ffd76e3205bdd8859e44ddf124220e1f2bae16438eec7c94214c4e3d9e37205
pillow @ file:///croot/pillow_1752524526621/work
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyQt6==6.9.1
PyQt6_sip @ file:///home/task_175751008926564/conda-bld/pyqt-split_1757511022658/work/pyqt_sip
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1752135852232/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests==2.32.5
safetensors==0.6.2
scikit-image @ file:///croot/scikit-image_1726737398091/work
scikit-learn @ file:///croot/scikit-learn_1737988764076/work
scipy @ file:///croot/scipy_1733756309941/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=a4372e07d982a377bda57ec71fd74ac403442571e1d717de47325a14dea5c278
sip @ file:///croot/sip_1756223371714/work
six @ file:///croot/six_1744271502820/work
sympy==1.14.0
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile @ file:///croot/tifffile_1695107451082/work
timm==1.0.20
tomli @ file:///croot/tomli_1753774587605/work
torch==2.8.0
torchvision==0.23.0
tornado @ file:///croot/tornado_1748956929273/work
tqdm @ file:///croot/tqdm_1738943501192/work
triton==3.4.0
typing_extensions==4.15.0
tzdata @ file:///croot/python-tzdata_1746123641790/work
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
urllib3==2.5.0
zipp @ file:///croot/zipp_1732630741423/work
